Always ip monitor link when setting negotiation parameters
The assumptions are that that ip monitor link reflects events that
have been seen by the ethernet driver which occurs after they are seen by
the phy. And that ethtool which reflects the state of the phy. So
ethtool should be run once ip monitor link indicates link change
events have propogated to the driver. Otherwise ethtool may indicate
the pre-link change state.

Signed-off-by: Simon Horman <horms@verge.net.au>